refactor(schematics): move setup tasks in separate folders
