+++ package.json
@@ -35,7 +35,7 @@
     "build:package": "npm-run-all -s clean:tmp:lib gulp build:ts build:ps build:inline-resources build:bundle clean:tmp",
     "build:ts": "tsc -p tsconfig.publish.json && ngc -p tsconfig.publish.json",
     "build:pst": "tsc -p tsconfig.publish.schematics.json",
-    "build:ps": "npm run build:pst",
+    "build:ps": "npm run build:pst && gulp inline-schematics-resources",
     "build:inline-resources": "gulp inline-resources",
     "build:bundle": "gulp bundle",
     "build:rename:dev": "gulp bundle:rename-dev",
