+++ scripts/gulp/tasks/inline-resources/inline-resources.ts
@@ -2,7 +2,7 @@ import { dest, src, task } from 'gulp';
 import { copyResources } from './copy-resources';
 import { BUILD_DIR, LIB_DIR } from '../config';
 
-task('inline-resources', () => {
+task('inline-resources', ['inline-schematics-resources'], () => {
   src([
     `${BUILD_DIR}/**/*.html`,
     `${BUILD_DIR}/**/*.css`,
@@ -11,6 +11,14 @@ task('inline-resources', () => {
     `${BUILD_DIR}/**/README.md`,
     `${BUILD_DIR}/**/package.json`,
     `${BUILD_DIR}/**/schematics/**/*.json`,
+  ])
+    .pipe(dest(LIB_DIR))
+    .on('end', () => copyResources(LIB_DIR));
+});
+
+task('inline-schematics-resources', () => {
+  src([
+    `./src/framework/**/schematics/**/*.json`,
   ])
     .pipe(dest(LIB_DIR))
     .on('end', () => copyResources(LIB_DIR));
