+++ src/framework/theme/schematics/ng-add/register-theme/theme-content.ts
@@ -5,26 +5,22 @@
  */
 
 export function createThemeContent(themeName: string): string {
-  return `
-@import '~@nebular/theme/styles/theming';
+  return `@import '~@nebular/theme/styles/theming';
 @import '~@nebular/theme/styles/themes/${themeName}';
 
 $nb-themes: nb-register-theme((
   // add your variables here like:
   // color-bg: #4ca6ff,
 ), ${themeName}, ${themeName});
-
 `;
 }
 
-export const stylesContent =  `
-@import 'themes';
+export const stylesContent =  `@import 'themes';
 
 @import '~@nebular/theme/styles/globals';
 
 @include nb-install() {
   @include nb-theme-global();
 };
-
 `;
